empty table fix

+++ Week-12/Day-2/fin-transaction/src/components/TransactionList.js
@@ -8,29 +8,36 @@ const TransactionList = (props) => {
   console.log(list)
   
   const dispatch = useDispatch();
-
-  return (
-    <div style={{display:'inline-block',width:'60%', verticalAlign: 'top'}}>
-      <h1>Transaction List</h1>
-      <table>
-      <tbody>
-      {
-        list.map((item, indx) => {
-          return(
-            <tr key={indx}>
-              <td> {item.accountNumber || '-'}</td>
-              <td> {item.FSC || '-'} </td>
-              <td> {item.accHolderName || '-'}</td>
-              <td> {item.amount || '-'} </td>
-              <td> <button onClick={() => dispatch({type:DETAIL, payload: item}) }>UPDATE</button></td>
-              <td> <button onClick={() => dispatch({type:DELETE, payload: String(indx)}) }>DELETE</button></td>
-            </tr>
+  if (list.length) {
+    return (
+      <div style={{display:'inline-block',width:'60%', verticalAlign: 'top'}}>
+        <h1>Transaction list</h1>
+        <table>
+        <tbody>
+        {
+          list.map((item, indx) => {
+            return(
+              <tr key={indx}>
+                <td> {item.accountNumber || '-'}</td>
+                <td> {item.FSC || '-'} </td>
+                <td> {item.accHolderName || '-'}</td>
+                <td> {item.amount || '-'} </td>
+                <td> <button onClick={() => dispatch({type:DETAIL, payload: item}) }>UPDATE</button></td>
+                <td> <button onClick={() => dispatch({type:DELETE, payload: String(indx)}) }>DELETE</button></td>
+              </tr>
+            )
+          })
+        }
+        </tbody>
+        </table>
+      </div>
+    )
+  } else {
+  return (<div style={{display:'inline-block',width:'60%', verticalAlign: 'top'}}>
+             <h1>Transaction list</h1>
+           </div>
           )
-        })
-      }
-      </tbody>
-      </table>
-    </div>
-  )
+  }
 }
+
 export default TransactionList
